profile/cpp/central/src/androidTest/java/org/im97mori/ble/profile/cpp/central/db/CyclingPowerProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.cpp.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class CyclingPowerProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(CyclingPowerProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(CyclingPowerProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), CyclingPowerProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(CyclingPowerProfileBondedDatabaseHelper.PROFILE_NAME, CyclingPowerProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
